rm StateManager.applyWriteOpts
