+++ packages/core/src/state-management/state-manager.ts
@@ -39,29 +39,6 @@ export class StateManager {
   ) {}
 
   /**
-   * Apply options relating to authoring a new commit
-   *
-   * @param state$ - Running State
-   * @param opts - Initialization options (request anchor, publish to pubsub, etc.)
-   * @param opType - If we load, create or update a stream
-   * @private
-   */
-  async applyWriteOpts(
-    state$: RunningState,
-    opts: CreateOpts | UpdateOpts,
-    opType: OperationType
-  ): Promise<void> {
-    const anchor = (opts as any).anchor
-    const publish = (opts as any).publish
-    if (anchor) {
-      await this.anchor(state$, opts)
-    }
-    if (publish && opType !== OperationType.LOAD) {
-      this.internals.publishTip(state$)
-    }
-  }
-
-  /**
    * Request anchor for the latest stream state
    */
   async anchor(state$: RunningState, opts: AnchorOpts): Promise<Subscription> {
